more comments and done with experimenting with new worker thread.
The worker thread works as intended. Next is including matrix
multiplication and vector addition.

+++ scheduling/schedulers/asyncCompletionQueue/completionQueue.h
@@ -18,7 +18,7 @@
 
 struct CompletionEvent {
   Job *job;
-  int completionTime;
+  float completionTime;
 };
 
 class CompletionQueue {
